Refactoring - move from controller to service
