+++ openpkw-services/src/main/java/org/openpkw/qr/service/QrService.java
@@ -1,11 +1,14 @@
 package org.openpkw.qr.service;
 
+import org.openpkw.qr.dto.QrDTO;
+import org.openpkw.qr.dto.QrResultDTO;
+
 /**
  * Service for parsing and saving qr from mobile
  * @author Sebastian Pogorzelski
  */
 public interface QrService {
 
-
+    QrResultDTO saveResult(QrDTO qrDTO);
 
 }
